fix some bugs and modification code

+++ src/app/modules/LostItem/lostItem.service.ts
@@ -4,6 +4,8 @@ import { JwtPayload } from "jsonwebtoken";
 import { TReportLostItem } from "./lostItem.interface";
 import { prisma } from "../../utils/prisma";
 import AppError from "../../errors/AppError";
+import { pagination } from "../../utils/pagination";
+import { TMeta, TPaginationQuery } from "../../interface/pagination";
 
 const reportLostItem = async (payload: {
   user: JwtPayload;
@@ -32,17 +34,28 @@ const reportLostItem = async (payload: {
   return result;
 };
 
-const getMyLostItems = async (payload: { user: JwtPayload }) => {
+const getMyLostItems = async (payload: {
+  user: JwtPayload;
+  query: { pagination: TPaginationQuery };
+}) => {
+  const { page, limit, skip, sortBy, sortOrder } = pagination(
+    payload.query.pagination,
+  );
+
   const result = await prisma.lostItem.findMany({
     where: {
       userId: payload.user.id,
     },
+    orderBy: {
+      [sortBy]: sortOrder,
+    },
+    skip: skip,
+    take: limit,
     select: {
       id: true,
       title: true,
       category: true,
       brand: true,
-      images: true,
       lostDate: true,
       lostLocation: true,
       isFound: true,
@@ -52,14 +65,30 @@ const getMyLostItems = async (payload: { user: JwtPayload }) => {
     },
   });
 
-  return result;
+  const meta: TMeta = {
+    page,
+    limit,
+    total: result.length,
+  };
+
+  return { data: result, meta: meta };
 };
 
-const getAllLostItems = async () => {
+const getAllLostItems = async (payload: {
+  query: { pagination: TPaginationQuery };
+}) => {
+  const { page, limit, skip, sortBy, sortOrder } = pagination(
+    payload.query.pagination,
+  );
   const result = await prisma.lostItem.findMany({
     where: {
       isFound: false,
     },
+    orderBy: {
+      [sortBy]: sortOrder,
+    },
+    skip: skip,
+    take: limit,
     select: {
       id: true,
       title: true,
@@ -72,7 +101,13 @@ const getAllLostItems = async () => {
     },
   });
 
-  return result;
+  const meta: TMeta = {
+    page,
+    limit,
+    total: result.length,
+  };
+
+  return { data: result, meta: meta };
 };
 
 const getSingleLostItem = async (payload: { lostItemId: string }) => {
@@ -188,6 +223,10 @@ const deleteLostItem = async (payload: {
 
   if (!result) {
     throw new AppError(404, "Item not found!");
+  }
+
+  if (result.isFound) {
+    throw new AppError(400, "Found item can't be deleted");
   }
 
   await prisma.lostItem.delete({ where: { id: result.id } });
